cloudy-crocodiles/django/requirements-dev.txt
Django>=3.1,<3.2
django-cors-headers
ariadne>=0.12,<0.13
psycopg2-binary
pytest-cov
pytest-django
Pillow
PyJWT
whitenoise